cabinet/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
faker==16.8.1
unidecode==1.3.6
tqdm==4.64.1